+++ src/client/components/SearchInput.tsx
@@ -3,16 +3,17 @@ import { connect } from "react-redux";
 
 import { searchJobs } from "../redux/thunks";
 
-import { LocationOption } from "../types";
+import { LocationOption, RootState } from "../types";
 
 interface SearchInputProps {
   handleSearch: (search: string, locationOptions: LocationOption[]) => void;
   locationOptions: LocationOption[];
+  searchValue: string;
 }
 
 const SearchInput: React.SFC<SearchInputProps> = (props: SearchInputProps) => {
-  const { handleSearch, locationOptions } = props;
-  const [search, setSearch] = React.useState("");
+  const { handleSearch, locationOptions, searchValue } = props;
+  const [search, setSearch] = React.useState(searchValue);
 
   return (
     <div className="search__container__outer">
@@ -42,9 +43,13 @@ const SearchInput: React.SFC<SearchInputProps> = (props: SearchInputProps) => {
   );
 };
 
+const mapStateToProps = (state: RootState) => ({
+  searchValue: state.application.searchValue,
+});
+
 const mapDispatchToProps = (dispatch) => ({
   handleSearch: (search: string, locationOptions: LocationOption[]) =>
     dispatch(searchJobs(search, locationOptions)),
 });
 
-export default connect(null, mapDispatchToProps)(SearchInput);
+export default connect(mapStateToProps, mapDispatchToProps)(SearchInput);