+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Page description
 - Favicon
 - If `company_url` exists, company title on `Details` page is rendered as a link
+- `searchValue` is retained in `ApplicationState`
 
 ### Removed